KN-22930 removed unnecesary converter
